Internals

MultilayerGraphs.AbstractGraphOfGraphs

MultilayerGraphs.AbstractInterlayer

MultilayerGraphs.AbstractLayer

MultilayerGraphs.AbstractMultilayerEdge

MultilayerGraphs.AbstractMultilayerGraph

MultilayerGraphs.AbstractMultilayerVertex

MultilayerGraphs.AbstractVertex

MultilayerGraphs.DiGraphOfGraphs

MultilayerGraphs.DiGraphOfGraphs

MultilayerGraphs.GraphOfGraphs

MultilayerGraphs.GraphOfGraphs

MultilayerGraphs.Interlayer

MultilayerGraphs.Interlayer

MultilayerGraphs.Interlayer

MultilayerGraphs.IsWeighted

MultilayerGraphs.Layer

MultilayerGraphs.Layer

MultilayerGraphs.Layer

MultilayerGraphs.MultilayerDiGraph

MultilayerGraphs.MultilayerDiGraph

MultilayerGraphs.MultilayerDiGraph

MultilayerGraphs.MultilayerDiGraph

MultilayerGraphs.MultilayerDiGraph

MultilayerGraphs.MultilayerEdge

MultilayerGraphs.MultilayerGraph

MultilayerGraphs.MultilayerGraph

MultilayerGraphs.MultilayerGraph

MultilayerGraphs.MultilayerGraph

MultilayerGraphs.MultilayerGraph

MultilayerGraphs.MultilayerVertex

MultilayerGraphs.δ

MultilayerGraphs.δ

MultilayerGraphs.δ_1

MultilayerGraphs.δ_2

MultilayerGraphs.δ_3

MultilayerGraphs.δ_Ω

SimpleWeightedGraphs.SimpleWeightedDiGraph

SimpleWeightedGraphs.SimpleWeightedDiGraph

SimpleWeightedGraphs.SimpleWeightedDiGraph

SimpleWeightedGraphs.SimpleWeightedGraph

SimpleWeightedGraphs.SimpleWeightedGraph

SimpleWeightedGraphs.SimpleWeightedGraph

Base.:==

Base.:==

Base.:==

Base.eltype

Base.eltype

Base.eltype

Base.getindex

Base.getindex

Base.getproperty

Base.getproperty

Base.getproperty

Base.size

Graphs.LinAlg.adjacency_matrix

Graphs.LinAlg.adjacency_matrix

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.rem_edge!

Graphs.SimpleGraphs.rem_edge!

Graphs.SimpleGraphs.rem_edge!

Graphs.SimpleGraphs.rem_edge!

Graphs.dst

Graphs.edges

Graphs.edges

Graphs.edges

Graphs.edgetype

Graphs.edgetype

Graphs.edgetype

Graphs.eigenvector_centrality

Graphs.has_edge

Graphs.has_edge

Graphs.has_edge

Graphs.has_edge

Graphs.has_edge

Graphs.has_edge

Graphs.has_edge

Graphs.has_vertex

Graphs.has_vertex

Graphs.has_vertex

Graphs.inneighbors

Graphs.inneighbors

Graphs.inneighbors

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.modularity

Graphs.ne

Graphs.ne

Graphs.ne

Graphs.nv

Graphs.nv

Graphs.nv

Graphs.outneighbors

Graphs.outneighbors

Graphs.outneighbors

Graphs.vertices

Graphs.vertices

Graphs.vertices

MultilayerGraphs._add_layer!

MultilayerGraphs._specify_interlayer!

MultilayerGraphs.add_layer!

MultilayerGraphs.add_layer!

MultilayerGraphs.adjm_eltype

MultilayerGraphs.check_unique

MultilayerGraphs.degree_second_moment

MultilayerGraphs.degree_variance

MultilayerGraphs.find_trunc_dim

MultilayerGraphs.format_trunc_chis

MultilayerGraphs.get_common_type

MultilayerGraphs.get_concrete_subtypes

MultilayerGraphs.get_diagonal_adjacency_tensor

MultilayerGraphs.get_diagonal_elements

MultilayerGraphs.get_graph_of_layers

MultilayerGraphs.get_graph_of_layers

MultilayerGraphs.get_interlayer

MultilayerGraphs.get_layer

MultilayerGraphs.get_oom

MultilayerGraphs.get_overlay_monoplex_graph

MultilayerGraphs.get_overlay_monoplex_graph

MultilayerGraphs.get_projected_monoplex_graph

MultilayerGraphs.get_projected_monoplex_graph

MultilayerGraphs.get_subgraph

MultilayerGraphs.get_symmetric_interlayer

MultilayerGraphs.is_weighted

MultilayerGraphs.isapproxsymmetric

MultilayerGraphs.isapproxsymmetric

MultilayerGraphs.iscompletelyinitialized

MultilayerGraphs.mean_degree

MultilayerGraphs.multilayer_global_clustering_coefficient

MultilayerGraphs.multilayer_kronecker_delta

MultilayerGraphs.multilayer_weighted_global_clustering_coefficient

MultilayerGraphs.multiplex_interlayer

MultilayerGraphs.nIn

MultilayerGraphs.nl

MultilayerGraphs.nn

MultilayerGraphs.nodes

MultilayerGraphs.overlay_clustering_coefficient

MultilayerGraphs.specify_interlayer!

MultilayerGraphs.specify_interlayer!

MultilayerGraphs.specify_interlayer!

MultilayerGraphs.src

MultilayerGraphs.tensoreig

MultilayerGraphs.tensorsplit

MultilayerGraphs.tensorsvd

MultilayerGraphs.to_matrix

MultilayerGraphs.von_neumann_entropy

MultilayerGraphs.weight

MultilayerGraphs.AbstractGraphOfGraphs — Type

abstract type AbstractGraphOfGraphs{T <: AbstractGraph} <: AbstractGraph{T} end

An abstract type representing a graph of graphs.

MultilayerGraphs.AbstractInterlayer — Type

AbstractInterlayer{T,U,G}

An abstract type representing a generic Interlayer.

FIELDS

T: the node type;

U: the adjacency matrix/tensor eltype;

G: the underlying graph type.

MultilayerGraphs.AbstractLayer — Type

AbstractLayer{T,U,G}

An abstract type representing a generic Layer.

FIELDS

T: the node type;

U: the adjacency matrix/tensor eltype;

G: the underlying graph type.

MultilayerGraphs.AbstractMultilayerEdge — Type

AbstractMultilayerEdge{T} <: AbstractEdge{T}

An abstract type representing a MultilayerGraph edge.

It must have fields: src, dst, weight.

MultilayerGraphs.AbstractMultilayerGraph — Type

AbstractMultilayerGraph{T <: Integer, U <: Real} <: AbstractGraph{T}

An abstract type for multilayer graphs, it must contain the fields: adjacency_tensor, layers, Interlayers. It is a subtype of AbstractGraph and its concrete subtypes may extend Graphs.jl.

MultilayerGraphs.AbstractMultilayerVertex — Type

AbstractMultilayerVertex{T} <: AbstractVertex{T}

An abstract type representing an abstract MultilayerGraph vertex.

MultilayerGraphs.AbstractVertex — Type

AbstractVertex{T}

An abstract type representing a graph vertex.

MultilayerGraphs.DiGraphOfGraphs — Type

mutable struct DiGraphOfGraphs{T} <: AbstractGraphOfGraphs{T}

Default directed concrete implementation of AbstractGraphOfGraphs.

FIELDS

nodes::OrderedDict{Int64,T}: An OrderedDict that associates graphs (values) to Int64 nodes (keys).

graph::SimpleWeightedGraph{Int64,Float64}: The weighted directed graph representing the graph of graphs. Its Int64 nodes are associated to graphs via the nodes field.

MultilayerGraphs.DiGraphOfGraphs — Method

DiGraphOfGraphs(graphs::Vector{T}, adjacency_matrix::Matrix{Float64})

Outer constructor for DiGraphOfGraphs.

ARGUMENTS

graphs::Vector{T}: The vector of nodes of the graphs of graphs. Each of them will be assigned to an integer, from 1 to length(graphs), and their order is assumed to be the same as the rows and columns of the adjacency_matrix argument.

adjacency_matrix::Matrix{Float64}: The adjacency matrix of the graph of graphs. Its (i,j)-element is the weight of the link between graphs[i] and graphs[j].

MultilayerGraphs.GraphOfGraphs — Type

mutable struct GraphOfGraphs{T} <: AbstractGraphOfGraphs{T}

Default undirected concrete implementation of AbstractGraphOfGraphs.

FIELDS

nodes::OrderedDict{Int64,T}: An OrderedDict that associates graphs (values) to Int64 nodes (keys).

graph::SimpleWeightedGraph{Int64,Float64}: The weighted undirected graph representing the graph of graphs. Its Int64 nodes are associated to graphs via the nodes field.

MultilayerGraphs.GraphOfGraphs — Method

GraphOfGraphs(graphs::Vector{T}, adjacency_matrix::Matrix{Float64})

Outer constructor for GraphOfGraphs.

ARGUMENTS

graphs::Vector{T}: The vector of nodes of the graphs of graphs. Each of them will be assigned to an integer, from 1 to length(graphs), and their order is assumed to be the same as the rows and columns of the adjacency_matrix argument.

adjacency_matrix::Matrix{Float64}: The adjacency matrix of the graph of graphs. Its (i,j)-element is the weight of the link between graphs[i] and graphs[j].

MultilayerGraphs.Interlayer — Type

mutable struct Interlayer{G <: AbstractGraph}

Represents an interlayer in a Multilayer(Di)Graph.

FIELDS

nv::Int64 : the number of nodes;

name`::Symbol : name of the Interlayer;

layer_1::Symbol: name of one of the Layers connected by this Interlayer;

layer_2::Symbol: name of one of the Layers connected by this Interlayer;

graph_type`::Type{G}: type of the graph underlying the Interlayer;

edge_list::Tuple{Vararg{ <: MultilayerEdge{MultilayerVertex{T},U} }}: edge list for the Interlayer;

forbidden_vertices::Vector{MultilayerVertex{T}}: nodes of the MultilayerGraph that are not part of this Interlayer (they will be formally present in the Interlayer but it will be checked that they aren't adjacacent to any other node);

forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}: edges that are required not to exist in this Interlayer.

Constructors

Interlayer(name::Symbol,layer_1::Symbol,layer_2::Symbol, graph::G, forbidden_vertices::Vector{MultilayerVertex{T}}, forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}; U::Union{Type{ <: Real}, Nothing} = nothing  ) where { T <: Union{ <: Integer, AbstractVertex}, G <: AbstractGraph{T}}

Overridden inner constructor. Return an Interlayer between layer_1 and layer_2, whose underlying graph is graph. All Layers and Interlayers of a Multilayer(Di)Graph need to formally have the same vertices, but in real applications it may be that some vertices are excluded from som layers. Such vertices should be specified in forbiddes_vertices. Similarly for forbiddes_edges. This constructor (to which all the other eventually fall back to) will check that forbidden_vertices have no neighbors in graph, and that forbidden_edges actually correspond to zero entries in the adjacency matrix of graph. Also check that it is a proper bipartite graph.

julia> Interlayer(:interlayer, Set((:layer_1,:layer_2)), SimpleGraph(50,50))
Interlayer{SimpleGraph{Int64}}(:interlayer, Set([:layer_2, :layer_1]), {50, 50} undirected simple Int64 graph)

MultilayerGraphs.Interlayer — Method

Interlayer(nv::Int64, name::Symbol,layer_1::Symbol, layer_2::Symbol, graph_type::Type{G}, ne::Int64, forbidden_vertices::Vector{MultilayerVertex{T}}, forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}) where {T <: Union{ <: Integer, AbstractVertex}, G <: AbstractGraph{T}; !IsDirected{G}, IsWeighted{G}}

Random Interlayer.

ARGUMENTS

nv::Int64: number of vertices;

name::Symbol: name of the Interlayer;

layer_1::Symbol: The first Layer it connects;

layer_2::Symbol: The second Layer it connects;

graph_type::Type{G}: the underlying graph type;

ne::Int64: the number of edges;

forbidden_vertices::Vector{MultilayerVertex{T}}: list of vertices that are not considered present in the Interlayer;

forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}: list of edges whose existence is a priori excluded from the Interlayer.

MultilayerGraphs.Interlayer — Method

Interlayer(nv::Int64, name::Symbol,layer_1::Symbol, layer_2::Symbol, graph_type::Type{G}, edge_list::Tuple{Vararg{ <: MultilayerEdge{MultilayerVertex{T},U} }}, forbidden_vertices::Vector{MultilayerVertex{T}}, forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}) where {T <: Union{ <: Integer, AbstractVertex}, U <: Real, G <: AbstractGraph{T}; IsWeighted{G}}

Outer constructor for Interlayer. Return an Interlayer between layer_1 and layer_2, whose underlying graph type is graph_type. Edges are given via an edge list edge_list. All Layers and Interlayers of a Multilayer(Di)Graph need to formally have the same vertices, but in real applications it may be that some vertices are excluded from som layers. Such vertices should be specified in forbiddes_vertices. Similarly for forbiddes_edges.

ARGUMENTS

nv::Int64 : the number of nodes;

name`::Symbol : name of the Interlayer;

layer_1::Symbol: name of one of the Layers connected by this Interlayer;

layer_2::Symbol: name of one of the Layers connected by this Interlayer;

graph_type`::Type{G}: type of the graph underlying the Interlayer;

edge_list::Tuple{Vararg{ <: MultilayerEdge{MultilayerVertex{T},U} }}: edge list for the Interlayer;

forbidden_vertices::Vector{MultilayerVertex{T}}: nodes of the MultilayerGraph that are not part of this Interlayer (they will be formally present in the Interlayer but it will be checked that they aren't adjacent to any other node);

forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}: edges that are required not to exist in this Interlayer.

MultilayerGraphs.IsWeighted — Type

IsWeighted{X}

Trait that discerns between weighted and unweighted graphs.

MultilayerGraphs.Layer — Type

mutable struct Layer{T <: Integer, U <: Real, G <: AbstractGraph{T}} <: AbstractLayer{T,U,G}

Represents a layer in a Multilayer(Di)Graph.

FIELDS

name::Symbol: the name of the layer;

graph::G: underlying graph of the layer;

forbidden_vertices::Vector{MultilayerVertex{T}}: nodes of the MultilayerGraph that are not part of this Layer (they will be formally present in the Layer but it will be checked that they aren't adjacent to any other node);

forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}: edges that are required not to exist in this Layer.

CONSTRUCTORS

Layer(name::Symbol, graph::G, forbidden_vertices::Tuple{Vararg{T}}, forbidden_edges::Tuple{Vararg{NTuple{2, T}}};  U::Union{Type{ <: Real}, Nothing}  = nothing ) where {T,G <: AbstractGraph{T}}

Overridden inner constructor. Return an Layer whose underlying graph is graph. All Layers and Layers of a Multilayer(Di)Graph need to formally have the same nodes, but in real applications it may be that some vertices are excluded from some layers. Such vertices should be specified in forbiddes_vertices. Similarly for forbiddes_edges. This constructor (to which all the other eventually fall back to) will check that forbidden_vertices have no neighbors in graph, and that forbidden_edges actually correspond to zero entries in the adjacency matrix of graph.

Layer{T <: Integer, U <: Real, G <: AbstractGraph{T}} <: AbstractLayer{T,U,G}

Incomplete initialization, used to write type-stable functions

MultilayerGraphs.Layer — Method

Layer(name::Symbol, graph::G; U::Union{Type{ <: Real}, Nothing} = nothing) where {G <: AbstractGraph}

Return a Layer with name name and graph graph with no forbidden nodes or edges.

MultilayerGraphs.Layer — Method

Layer(nv::Int64, name::Symbol, graph_type::Type{G}, ne::Int64; U::Union{Type{ <: Real},Nothing} = nothing)  where {T <: Union{ <: Integer, AbstractVertex}, G <: AbstractGraph{T}}

Random Layer.

ARGUMENTS

nv::Int64: number of vertices;

name::Symbol: name of the layer;

graph_type::Type{G}: the underlying graph type;

ne::Int64: the number of edges;

MultilayerGraphs.MultilayerDiGraph — Type

MultilayerDiGraph{T, U} <: AbstractMultilayerDiGraph{T,U}

A concrete type that can represent a general directed multilayer graph.

FIELDS

adjacency_tensor::Array{U, 4}: the 4-dimensional tensor that encodes all (weighted) connections within the graph. adjacency_tensor[1,2,3,4] encodes the strength of the (directed or undirected) link between node 1 in layer 3 and node 2 in layer 4.

layers::OrderedDict{ Tuple{Int64,Int64}, Layer{T,U,G}}: the ordered dictionary containing all the layers of the multilayer graph. Their underlying graphs must be all directed.

interlayers::OrderedDict{ Tuple{Int64,Int64}, Interlayer{T,U}}: the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all directed.

MultilayerGraphs.MultilayerDiGraph — Method

MultilayerDiGraph(num_layers::Int64, n_nodes::Int64, min_edges::Int64, max_edges::Int64, graph_type::Type{<: AbstractGraph})

Return a random MultilayerDiGraph with num_layers layers, n_nodes nodes and each Layer and Interlayer has a random number of edges between min_edges and max_edges. Layers and Interlayers have parametric type graph_type.

MultilayerGraphs.MultilayerDiGraph — Method

MultilayerDiGraph(n_nodes::Int64, T::Type{ <: Number}, U::Type{ <: Number} )

Return a MultilayerDiGraph with n_nodes of type T nodes and an adjacency tensor eltype U. Use this constructor and then add Layers and Interlayers via the add_layer! and add_interlayer! methods.

MultilayerGraphs.MultilayerDiGraph — Method

MultilayerDiGraph(layers::Vector{ <: Layer{T,U }}, specified_interlayers::Vector{ <: Interlayer{T,U}};  default_interlayer::String  = "multiplex") where {T, U}

Construct a MultilayerDiGraph with layers given by layers. The interlayers will be constructed by default according to default_interlayer (only "multiplex" is allowed), except for those specified in specified_interlayers.

MultilayerGraphs.MultilayerDiGraph — Method

MultilayerDiGraph(layers::Vector{ <: Layer{T,U}};  default_interlayer::String  = "multiplex") where {T,U}

Construct a MultilayerDiGraph with layers layers and all interlayers of type default_interlayer (only "multiplex" is allowed).

MultilayerGraphs.MultilayerEdge — Type

struct MultilayerEdge{ T <: MultilayerVertex, U <: Union{ <: Real, Nothing}} <: AbstractMultilayerEdge{T}

Default concrete subtype of AbstractMultilayerEdge.

FIELDS

src::T: the source vertex of the edge ;

dst::T: the destination vertex of the edge;

weight::U: the edge weight.

CONSTRUCTORS

MultilayerEdge(src::T, dst::T, weight::U) where { T <: MultilayerVertex, U <: Union{ <: Real, Nothing}}

Default constructor.

MultilayerEdge(src::T, dst::T) where {T <: MultilayerVertex}

Unweighted edge.

MultilayerGraphs.MultilayerGraph — Type

MultilayerGraph{T, U, G <: AbstractGraph{T}} <: AbstractMultilayerGraph{T,U}

A concrete type that can represent a general multilayer graph.

FIELDS

adjacency_tensor::Array{U, 4}: the 4-dimensional tensor that encodes all (weighted) connections within the graph. adjacency_tensor[1,2,3,4] encodes the strength of the (directed or undirected) link between node 1 in layer 3 and node 2 in layer 4.

layers::OrderedDict{ Tuple{Int64,Int64}, Layer{T,U,G}}: the ordered dictionary containing all the layers of the multilayer graph. Their underlying graphs must be all undirected.

interlayers::OrderedDict{ Tuple{Int64,Int64}, Interlayer{T,U}}: the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all undirected.

MultilayerGraphs.MultilayerGraph — Method

MultilayerGraph(num_layers::Int64, n_nodes::Int64, min_edges::Int64, max_edges::Int64, graph_type::Type{<: AbstractGraph})

Return a random MultilayerGraph with num_layers layers, n_nodes nodes and each Layer and Interlayer has a random number of edges between min_edges and max_edges. Layers and Interlayers have parametric type graph_type.

MultilayerGraphs.MultilayerGraph — Method

MultilayerGraph(n_nodes::Int64, T::Type{ <: Number}, U::Type{ <: Number} )

Return a MultilayerGraph with n_nodes of type T nodes and an adjacency tensor eltype U. Use this constructor and then add Layers and Interlayers via the add_layer! and add_interlayer! methods.

MultilayerGraphs.MultilayerGraph — Method

MultilayerGraph(layers::Vector{ <: Layer{T,U }}, specified_interlayers::Vector{ <: Interlayer{T,U}};  default_interlayer::String  = "multiplex") where {T, U}

Construct a MultilayerDiGraph with layers given by layers. The interlayers will be constructed by default according to default_interlayer (only "multiplex" is allowed), except for those specified in specified_interlayers.

MultilayerGraphs.MultilayerGraph — Method

MultilayerGraph(layers::Vector{ <: Layer{T,U}};  default_interlayer::String  = "multiplex") where {T,U}

Construct a MultilayerDiGraph with layers layers and all interlayers of type default_interlayer (only "multiplex" is allowed).

MultilayerGraphs.MultilayerVertex — Type

MultilayerVertex{T <: Integer} <: AbstractMultilayerVertex{T}

A struct representing a vertex of a MultilayerGraph.

FIELDS

vertex::T : the node that the MultilayerVertex represents;

layer::Symbol: the layer the MultilayerVertex belongs to.

MultilayerGraphs.δ — Type

mutable struct δ{T} <: AbstractVector{T}

The kronecker delta.

FIELDS

N::Int64: the number of dimensions;

representation::Matrix{Int64}: the matrix representing the kronecker delta;

T: the return type when called ad δ[i,j].

CONSTRUCTORS

δ{T}(N::Int64) where {T <: Number}

Inner constructor that only requires N and the eltype.

MultilayerGraphs.δ — Method

δ(N::Int64)

Outer constructor that only requires N

MultilayerGraphs.δ_1 — Type

struct δ_1{T<: Number}

The δ_1 from De Domenico et al. (2013). Evaluate it via the notation [i,j].

FIELDS

N:Int64: the dimensionality of δ_1;

T: the return type.

CONSTRUCTORS

δ_1{T<: Number}(N::Int64)

MultilayerGraphs.δ_2 — Type

struct δ_2{T<: Number}

The δ_2 from De Domenico et al. (2013). Evaluate it via the notation [i,j].

FIELDS

N:Int64: the dimensionality of δ_2;

T: the return type.

CONSTRUCTORS

δ_2{T<: Number}(N::Int64)

MultilayerGraphs.δ_3 — Type

struct δ_3{T<: Number}

The δ_3 from De Domenico et al. (2013). Evaluate it via the notation [i,j].

FIELDS

N:Int64: the dimensionality of δ_3;

T: the return type.

CONSTRUCTORS

δ_3{T<: Number}(N::Int64)

MultilayerGraphs.δ_Ω — Type

δ_Ω{T} <: AbstractVector{T}

Struct that represents the δ_Ω defined in De Domenico et al. (2013).

FIELDS

δ_1::δ_1{T}: Instance of δ_1;

δ_2::δ_2{T}: Instance of δ_2;

δ_3::δ_3{T}: Instance of δ_3;

N::Int64 : Maximum index (number of layers);

representation::Array{Int64,4}: Multidimensional-array representation of δ_Ω.

SimpleWeightedGraphs.SimpleWeightedDiGraph — Method

SimpleWeightedDiGraph(n_vertices::Integer, n_edges::Integer; T::Type = Int64, U::Type = Float64)

Random SimpleWeightedDiGraph with n_vertices vertices and n_edges edges, vertex type T and adjacency matrix eltype U. Edge weights are uniformly extracted between 0 and 1.

SimpleWeightedGraphs.SimpleWeightedDiGraph — Method

SimpleWeightedGraphs.SimpleWeightedDiGraph{T}(n_vertices::Integer, n_edges::Integer; U::Type = Float64) where {T}

Random SimpleWeightedDiGraph with n_vertices vertices and n_edges edges, vertex type T and adjacency matrix eltype U. Edge weights are uniformly extracted between 0 and 1.

SimpleWeightedGraphs.SimpleWeightedDiGraph — Method

SimpleWeightedDiGraph{T,U}(n_vertices::Integer, n_edges::Integer)

Random SimpleWeightedDiGraph with n_vertices vertices and n_edges edges, vertex type T and adjacency matrix eltype U. Edge weights are uniformly extracted between 0 and 1.

SimpleWeightedGraphs.SimpleWeightedGraph — Method

SimpleWeightedGraph(n_vertices::Integer, n_edges::Integer; T::Type = Int64, U::Type = Float64)

Random SimpleWeightedGraph with n_vertices vertices and n_edges edges, vertex type T and adjacency matrix eltype U. Edge weights are uniformly extracted between 0 and 1.

SimpleWeightedGraphs.SimpleWeightedGraph — Method

SimpleWeightedGraph{T}(n_vertices::Integer, n_edges::Integer; U::Type = Float64) where { T }

Random SimpleWeightedGraph with n_vertices vertices and n_edges edges, vertex type T and adjacency matrix eltype U. Edge weights are uniformly extracted between 0 and 1.

SimpleWeightedGraphs.SimpleWeightedGraph — Method

SimpleWeightedGraph{T,U}(n_vertices::Integer, n_edges::Integer) where { T, U }

Random SimpleWeightedGraph with n_vertices vertices and n_edges edges, vertex type T and adjacency matrix eltype U. Edge weights are uniformly extracted between 0 and 1.

Base.:== — Method

Base.(==)(x::AbstractMultilayerGraph, y::AbstractMultilayerGraph)

Overload equality for AbstractMultilayerGraphs.

Base.:== — Method

Base.(==)(x::Interlayer, y::Interlayer)

Overload equality for Interlayers.

Base.:== — Method

Base.(==)(x::Layer, y::Layer)

Overload equality for Layers.

Base.eltype — Method

Base.eltype(interlayer::In) where { In <: Interlayer}

Return the vertex type of interlayer.

Base.eltype — Method

Base.eltype(layer::L) where { L <: Layer}

Return the vertex type of layer.

Base.eltype — Method

Base.eltype(mg::M) where {M <: AbstractMultilayerGraph}

Return the vertex type of mg.

Base.getindex — Method

getindex(od::O, key::Int64) where {O <: OrderedDict}

Return the keyth pair of od, following od's order.

Base.getindex — Method

getindex(d::δ{T}, i::Int) where T

The getindex called by OMEinsum

Base.getproperty — Method

Base.getproperty(interlayer::In, f::Symbol) where { In <: Interlayer}

Base.getproperty — Method

Base.getproperty(layer::L, f::Symbol) where { L <: layer}

Base.getproperty — Method

Base.getproperty(mg::M, f::Symbol) where { M <: AbstractMultilayerGraph }

Base.size — Method

size(d::δ)

Override required by OMEinsum

Graphs.LinAlg.adjacency_matrix — Method

adjacency_matrix(interlayer::In) where { T,U, G <: AbstractGraph{T}, In <:Interlayer{T,U,G}}

Return the adjacency matrix of interlayer.graph, with the eltype converted to U.

Graphs.LinAlg.adjacency_matrix — Method

adjacency_matrix(layer::L) where { T,U, G <: AbstractGraph{T}, L <:Layer{T,U,G}}

Return the adjacency matrix of layer.graph, with the eltype converted to U.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(interlayer::In, me::E) where {T,U,G, In <: Interlayer{T,U,G}, E <: MultilayerEdge{MultilayerVertex{T},Nothing}}

Add unweighted edge me to interlayer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(interlayer::In, me::E) where {T,U <: Real,G, In <: Interlayer{T,U,G}, E <: MultilayerEdge{MultilayerVertex{T},U}}

Add unweighted edge me to interlayer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(layer::L, me::E) where {T,U,G<: AbstractGraph{T}, L <: Layer{T,U,G}, E <: MultilayerEdge{MultilayerVertex{T},Nothing}; !IsWeighted{G}}

Add unweighted edge me to layer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(layer::L, me::E) where {T,U <: Real,G <: AbstractGraph{T} , L <: Layer{T,U,G}, E <: MultilayerEdge{MultilayerVertex{T},U}; IsWeighted{G}}

Add unweighted edge me to layer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, me::E) where { T, U, M <: MultilayerDiGraph{T,U}, E <: MultilayerEdge{MultilayerVertex{T},Nothing}}

Add unweighted edge me to mg.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, me::E) where { T, U, M <: MultilayerGraph{T,U}, E <: MultilayerEdge{MultilayerVertex{T},Nothing}}

Add unweighted edge me to mg.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, me::E) where { T, U <: Real, M <: MultilayerDiGraph{T,U}, E <: MultilayerEdge{MultilayerVertex{T},U}}

Add weighted edge me to mg.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, me::E) where { T, U <: Real, M <: MultilayerGraph{T,U}, E <: MultilayerEdge{MultilayerVertex{T},U}}

Add weighted edge me to mg.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(interlayer::In,src::V, dst::V, weight::U) where { T, U <: Real, G <: AbstractGraph{T}, In <: Interlayer{T,U,G}, V <: MultilayerVertex{T}; IsWeighted{G}}

Add edge from src to dst with weight weight to interlayer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(interlayer::In,src::V, dst::V) where { T, U <: Real, G <: AbstractGraph{T}, In <: Interlayer{T,U,G}, V <: MultilayerVertex{T}; !IsWeighted{G}}

Add edge from src to dst with weight weight to interlayer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(layer::L,src::V, dst::V, weight::U) where { T, U <: Real, G <: AbstractGraph{T}, L <: Layer{T,U,G}, V <: MultilayerVertex{T}; IsWeighted{G}}

Add edge from src to dst with weight weight to layer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(layer::L, src::V, dst::V) where { T, U <: Real, G <: AbstractGraph{T}, L <: Layer{T,U,G}, V <: MultilayerVertex{T}; !IsWeighted{G}}

Add edge from src to dst with weight weight to layer.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, src::V, dst::V, weight::U) where { T, U <: Real, M <: MultilayerDiGraph{T,U}, V <: MultilayerVertex{T}}

Add edge from src to dst with weight weight to mg.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, src::V, dst::V, weight::U) where { T, U <: Real, M <: MultilayerGraph{T,U}, V <: MultilayerVertex{T}}

Add edge from src to dst with weight weight to mg.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, src::V, dst::V, weight::U) where { T, U <: Real, M <: MultilayerDiGraph{T,U}, V <: MultilayerVertex{T}}

Add edge from src to dst with weight one(T) to mg.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(mg::M, src::V, dst::V, weight::U) where { T, U <: Real, M <: MultilayerGraph{T,U}, V <: MultilayerVertex{T}}

Add edge from src to dst with weight one(U) to mg.

Graphs.SimpleGraphs.rem_edge! — Method

rem_edge!(interlayer::In, src::V, dst::V) where { T, U, In <: Interlayer{T,U}, V <: MultilayerVertex{T}}

Remove edge from src to dst in interlayer.

Graphs.SimpleGraphs.rem_edge! — Method

rem_edge!(layer::L, src::V, dst::V) where { T, U, G<: AbstractGraph{T}, L <:Layer{T,U,G}, V <: MultilayerVertex{T}}

Remove edge from src to dst in layer.

Graphs.SimpleGraphs.rem_edge! — Method

rem_edge!(mg::M, src::V, dst::V) where { T, U, M <: MultilayerDiGraph{T,U}, V <: MultilayerVertex{T}}

Remove edge from src to dst in mg.

Graphs.SimpleGraphs.rem_edge! — Method

rem_edge!(mg::M, V1::V, V2::V) where { T, U, M <: MultilayerGraph{T,U}, V <: MultilayerVertex{T}}

Remove edge from src to dst in mg.

Graphs.dst — Method

dst(e::AbstractMultilayerEdge)

Graphs.edges — Method

Graphs.edges(interlayer::In) where {T,U,G <: AbstractGraph{T}, In <: Interlayer{T,U,G}}

Return an iterator over all the edges of interlayer.

Graphs.edges — Method

Graphs.edges(layer::L) where {T,U,G <: AbstractGraph{T}, L <: Layer{T,U,G}}

Return an iterator over all the edges of layer.

Graphs.edges — Method

edges(mg::M) where {M <: AbstractMultilayerDiGraph}

Return an iterator over all the edges of mg. The iterators first loops over all layers' edges (in the order they are given in mg.layers), then over all interlayers' edges (in the order they are given in mg.interlayers).

Graphs.edgetype — Method

edgetype(interlayer::In) where {T,U,G,In <: Interlayer{T,U,G} }

Return the edge type for interlayer.

Graphs.edgetype — Method

edgetype(layer::L) where {T,U,G,L <: Layer{T,U,G} }

Return the edge type for layer.

Graphs.edgetype — Method

edgetype(mg::M) where {M <: AbstractMultilayerGraph}

Return the edge type for mg.

Graphs.eigenvector_centrality — Method

eigenvector_centrality(mg::M; norm::String = "1", tol::Float64 = 1e-6, maxiter::Int64 = 2000) where {T, U, M <: AbstractMultilayerGraph{T, U}}

Calculate the eigenvector centrality of mg via an iterative algorithm. The norm parameter may be "1" or "n", and respectively the eigenvector centrality will be normalized to 1 or further divided by the number of nodes of mg. The tol parameter terminates the approximation when two consecutive iteration differ by no more than tol. The maxiters parameter terminates the algorithm when it goes beyond maxiters iterations.

The returned values are: the eigenvector centrality and the relative error at each algorithm iteration, that is, the summed absolute values of the componentwise differences between the centrality computed at the current iteration minus the centrality computed at the previous iteration.

Note: in the limit case of a monoplex graph, this function outputs a eigenvector centrality vector that is a multiple of the one outputted by Graphs.jl's eigenvector_centrality. It is currently under investigation.

Graphs.has_edge — Method

has_edge(interlayer::In, s::MultilayerVertex{T}, d::MultilayerVertex{T}) where { T,U,G, In <: Interlayer{T,U,G}}

Return true if there is an edge between s and d, false otherwise.

Graphs.has_edge — Method

has_edge(interlayer::In,  s::MultilayerVertex{T}, d::MultilayerVertex{T}, weight::U) where { T,U,G, In <: Interlayer{T,U,G}; IsDirected{G}

Return true if there is an edge between s and d with weight weight, false otherwise.

Graphs.has_edge — Method

has_edge(interlayer::In,  s::MultilayerVertex{T}, d::MultilayerVertex{T}, weight::U) where { T,U,G, In <: Interlayer{T,U,G}; !IsDirected{G}}

Return `true` if there is an edge between `s` and `d` with weight `weight`, `false` otherwise.

Graphs.has_edge — Method

has_edge(layer::L, s::MultilayerVertex{T}, d::MultilayerVertex{T}) where { T,U,G, L <: Layer{T,U,G}}

Return true if there is an edge between s and d, false otherwise.

Graphs.has_edge — Method

has_edge(layer::L,  s::MultilayerVertex{T}, d::MultilayerVertex{T}, weight::U) where { T,U,G, L <: Layer{T,U,G}; IsDirected{G}}

Return true if there is an edge between s and d with weight weight, false otherwise.

Graphs.has_edge — Method

has_edge(layer::L,  s::MultilayerVertex{T}, d::MultilayerVertex{T}, weight::U) where { T,U,G, L <: Layer{T,U,G}; IsDirected{G}}

Return true if there is an edge between s and d with weight weight, false otherwise.

Graphs.has_edge — Method

has_edge(mg::M,s::T,d::T) where {M <: AbstractMultilayerGraph{T} } where { T <: Integer}

Return true if there is an edge between s and d, false otherwise. In case s and d belong to two different layers, but the link is found only in one of the two corresponding interlayers between them, throws an error.

Graphs.has_vertex — Method

has_vertex(interlayer::In, v::MultilayerVertex{T}) where { T,U,G, In <: Interlayer{T,U,G}}

Return true if v is a vertex of interlayer.

Graphs.has_vertex — Method

has_vertex(layer::L, v::Integer) where { L <: Layer}

Return true if v is a vertex of layer.

Graphs.has_vertex — Method

has_vertex(mg::M, v::T) where {M <: AbstractMultilayerGraph{T} } where { T <: Integer}

Return true if v is a vertex of mg.

Graphs.inneighbors — Method

inneighbors(interlayer::In, mv::MultilayerVertex{T}) where {T,U,G, In <: Interlayer{T,U,G}}

Return the list of inneighbors of v within interlayer.

Graphs.inneighbors — Method

inneighbors(layer::L, mv::MultilayerVertex{T}) where {T,U,G, L <: Layer{T,U,G}}

Return the list of inneighbors of v within layer.

Graphs.inneighbors — Method

inneighbors(mg::M, v::T) where {M <: AbstractMultilayerGraph{T} } where { T <: Integer}

Return the list of inneighbors of v within mg, looping first over all layers (in the order they are given in mg.layers), then over all interlayers (in the order they are given in mg.interlayers).

Graphs.is_directed — Method

is_directed(interlayer::In) where { In <: Interlayer}

Returns true if interlayer is directed, false otherwise.

Graphs.is_directed — Method

is_directed(layer::L) where { L <: Layer}

Returns true if layer is directed, false otherwise.

Graphs.is_directed — Method

is_directed(mg::M) where { M <: Type{ <: MultilayerDiGraph}}

Returns true if mg is directed, false otherwise.

Graphs.is_directed — Method

is_directed(m::M) where { M <: Type{ <: MultilayerGraph}}

Returns true if m is directed, false otherwise.

Graphs.is_directed — Method

is_directed(mg::M) where { M <: MultilayerDiGraph}

Returns true if mg is directed, false otherwise.

Graphs.is_directed — Method

is_directed(m::M) where { M <: MultilayerGraph}

Returns true if m is directed, false otherwise.

Graphs.is_directed — Method

is_directed(interlayer_type::Type{In}) where {In <: Interlayer}

Returns true if interlayer_type is directed, false otherwise.

Graphs.is_directed — Method

is_directed(::Type{L}) where {T,U,G,L <: Layer{T,U,G}}

Returns true if layer_type is directed, false otherwise.

Graphs.modularity — Method

modularity(mg::M, c::Matrix{Int64}; null_model::Union{String,Array{U,4}} = "degree") where {T, U, M <: AbstractMultilayerGraph{T,U}}

Calculate the modularity of mg, as shown in De Domenico et al. (2013).

Graphs.ne — Method

ne(interlayer::In) where { In <: Interlayer}

Return the number of edges in interlayer.

Graphs.ne — Method

ne(layer::L) where { L <: Layer}

Return the number of edges in layer.

Graphs.ne — Method

ne(mg::M) where {M <: AbstractMultilayerGraph }

Return the number of edges in mg.

Graphs.nv — Method

nv(interlayer::In) where { In <: Interlayer}

Return the number of vertices in interlayer.

Graphs.nv — Method

nv(layer::L) where { L <: Layer}

Return the number of vertices in layer.

Graphs.nv — Method

nv(mg::M) where {M <: AbstractMultilayerGraph }

Return the number of vertices in mg.

Graphs.outneighbors — Method

outneighbors(interlayer::In, v::T) where {In <: Interlayer{T} } where { T <: Integer}

Return the list of outneighbors of v within interlayer, looping first over all layers (in the order they are given in interlayer.layers), then over all interlayers (in the order they are given in interlayer.interlayers).

Graphs.outneighbors — Method

outneighbors(layer::L, mv::MultilayerVertex{T}) where {T,U,G, L <: Layer{T,U,G}} =

Return the list of outneighbors of v within layer.

Graphs.outneighbors — Method

outneighbors(mg::M, v::T) where {M <: AbstractMultilayerGraph{T} } where { T <: Integer}

Return the list of outneighbors of v within mg, looping first over all layers (in the order they are given in mg.layers), then over all interlayers (in the order they are given in mg.interlayers).

Graphs.vertices — Method

vertices(interlayer::In) where {In <: Interlayer{ <: Integer, <: AbstractSimpleGraph}}

Return the collection of the vertices of interlayer.

Graphs.vertices — Method

vertices(layer::L) where { L <: Layer}

Return the collection of the vertices of layer.

Graphs.vertices — Method

vertices(mg::M) where {M <: AbstractMultilayerGraph{ <: Integer, <: AbstractSimpleGraph}}

Return the collection of the vertices of mg.

MultilayerGraphs._add_layer! — Method

_add_layer!(mg::M,new_layer::L; new_default_interlayers_type::H) where { T, U, G <: AbstractGraph{T}, H <: AbstractGraph{T}, M <: AbstractMultilayerGraph{T, U}, L <: Layer{T,U,G}

Internal function. It is called by the add_layer! API functions, which needs to specify the default interlayer graph type (SimpleGraph for undirected multilayer graphs, SimpleDiGraph for directed multilayer graphs).

MultilayerGraphs._specify_interlayer! — Method

_specify_interlayer!(mg::M, new_interlayer::In; symmetric_interlayer_name::String) where { T, U, G <: AbstractGraph{T}, M <: AbstractMultilayerGraph{T, U}, In <: Interlayer{T,U,G}}

Internal function. It is called by the specify_interlayer! API functions.

MultilayerGraphs.add_layer! — Method

add_layer!(mg::M,layer::L; interlayers_type = "multiplex") where { T, undef, M <: MultilayerDiGraph{T, G, U}, L <: Layer{T,U,G}}

Add layer layer to mg. Also add Interlayers of type interlayers_type (can only be "multiplex") between the new layer and all the other ones.

MultilayerGraphs.add_layer! — Method

add_layer!(mg::M,layer::L; interlayers_type = "multiplex") where { T, U, G<: AbstractGraph{T}, M <: MultilayerGraph{T, U}, L <: Layer{T,U,G}}

Add layer layer to mg. Also add Interlayers of type interlayers_type (can only be "multiplex") between the new layer and all the other ones.

MultilayerGraphs.adjm_eltype — Method

adjm_eltype(mg::M) where {M <: AbstractMultilayerGraph}

Return the vertex type of mg.

MultilayerGraphs.check_unique — Method

check_unique(vec::Union{Missing, <: Vector, <: Tuple})

Return true if all elements in vec are unique or if ismissing(vec), else return false.

MultilayerGraphs.degree_second_moment — Method

degree_second_moment(mg::M) where { M <: AbstractMultilayerGraph}

Calculate the second moment of the degree sequence of mg.

MultilayerGraphs.degree_variance — Method

degree_variance(mg::M) where { M <: AbstractMultilayerGraph}

Return the variance of the degree sequence of mg.

MultilayerGraphs.find_trunc_dim — Function

Finds the dimension to which v should be truncated, and the error caused in this truncation. See documentation for tensorsvd for the meaning of the different arguments.

MultilayerGraphs.format_trunc_chis — Method

Format the bond dimensions listed in chis to a standard format.

MultilayerGraphs.get_common_type — Method

get_common_type(types::Vector{<: DataType})

Returns the minimum common supertype of types.

MultilayerGraphs.get_concrete_subtypes — Method

get_concrete_subtypes(type::Type)

Return an array of all concrete subtypes of type (which should be abstract).

MultilayerGraphs.get_diagonal_adjacency_tensor — Method

get_diagonal_adjacency_tensor(arr,dims)

Returns a tensor whose size is dims and has arr on the diagonal.

MultilayerGraphs.get_diagonal_elements — Method

get_diagonal_elements(arr::Array{T,N}) where {T,N}

MultilayerGraphs.get_graph_of_layers — Method

get_graph_of_layers(mg::M) where {M <: MultilayerDiGraph}

Get a DiGraphOfGraph of the layers of mg. the weight of each edge between layers are obtained by summing all edge weights in the corresponding interlayer. See De Domenico et al. (2013).

MultilayerGraphs.get_graph_of_layers — Method

get_graph_of_layers(mg::M) where {M <: MultilayerGraph}

Get a DiGraphOfGraph of the layers of mg. the weight of each edge between layers are obtained by summing all edge weights in the corresponding interlayer. See De Domenico et al. (2013).

MultilayerGraphs.get_interlayer — Method

get_interlayer(mg::M, layer_1::Symbol, layer_2::Symbol) where {M <: AbstractMultilayerGraph}

Return (idxs_cart,interlayer) where interlayer is the Interlayer between layer_1 and layer_2 and idxs_cart is the CartesianIndex such that mg.adjacency_tensor[:,:,idxs_cart] is the adjacency matrix of interlayer.

MultilayerGraphs.get_layer — Method

get_layer(mg::M, layer_name::Symbol) where {T, U, M <: AbstractMultilayerGraph{T, U}}

Return (idxs_cart,layer) where layer_name is the name of the Layer being queried and idxs_cart is the CartesianIndex such that mg.adjacency_tensor[:,:,idxs_cart] is the adjacency matrix of such layer.

MultilayerGraphs.get_oom — Method

get_oom(x::Number)

Get the order of magnitude of x.

MultilayerGraphs.get_overlay_monoplex_graph — Method

get_overlay_monoplex_graph(mg::M) where {M<: MultilayerDiGraph}

Get overlay monoplex graph (i.e. the graph that has the same nodes as mg but the link between node i and j has weight equal to the sum of all edges weights between the various vertices representing i and j in mg, accounting for both layers and interlayers). See De Domenico et al. (2013).

MultilayerGraphs.get_overlay_monoplex_graph — Method

get_overlay_monoplex_graph(mg::M) where {M<: MultilayerGraph}

Get overlay monoplex graph (i.e. the graph that has the same nodes as mg but the link between node i and j has weight equal to the sum of all edges weights between the various vertices representing i and j in mg, accounting for both layers and interlayers). See De Domenico et al. (2013).

MultilayerGraphs.get_projected_monoplex_graph — Method

get_projected_monoplex_graph(mg::M) where {M<: MultilayerDiGraph}

Get projected monoplex graph (i.e. the graph that as the same nodes as mg but the link between node i and j has weight equal to the sum of all edges weights between the various vertices representing i and j in mg, accounting for both layers and interlayers). See De Domenico et al. (2013).

MultilayerGraphs.get_projected_monoplex_graph — Method

get_projected_monoplex_graph(mg::M) where {M<: MultilayerGraph}

Get projected monoplex graph (i.e. that graph that as the same nodes as mg but the link between node i and j has weight equal to the sum of all edges weights between the various vertices representing i and j in mg, accounting for both layers and interlayers). See De Domenico et al. (2013).

MultilayerGraphs.get_subgraph — Method

get_subgraph(mg::M, names...) where {M <: AbstractMultilayerGraph}

If length(names) == 1, then this is equivalent to calling get_layer(mg,names[1]) else if length(names) == 2, this is equivalent to calling get_interlayer(mg,names[1],names[2]). Throws an ErrorException in all other cases.

MultilayerGraphs.get_symmetric_interlayer — Method

get_symmetric_interlayer(interlayer::In; symmetric_interlayer_name::String) where{T,U,G, In <: Interlayer{T,U,G}}

Return the Interlayer corresponding to interlayer where layer_1 and layer_2 are swapped. Its name will be symmetric_interlayer_name (defaults to interlayer_(interlayer.layer_2)_(interlayer.layer_1) ).

MultilayerGraphs.is_weighted — Method

is_weighted(g::G) where { G <: AbstractGraph}

Check whether g is weighted.

MultilayerGraphs.isapproxsymmetric — Method

isapproxsymmetric(A::Matrix{T}) where {T <: Real }

Check whether A is symmetric (within zero(T)).

MultilayerGraphs.isapproxsymmetric — Method

isapproxsymmetric(A::Matrix{T}) where {T <: Real }

"Check whether A is approximately symmetric (within eps(T)).

MultilayerGraphs.iscompletelyinitialized — Method

iscompletelyinitialized(obj::Any)

Check whether obj is completely initialized.

MultilayerGraphs.mean_degree — Method

mean_degree(mg::M) where { M <: AbstractMultilayerGraph}

Return the mean of the degree sequence of mg.

MultilayerGraphs.multilayer_global_clustering_coefficient — Method

multilayer_clustering_coefficient(mg::M, norm_factor::Union{Float64, Symbol} = :max) where {M <: AbstractMultilayerGraph}

Return the complete multilayer global clustering coefficient, equal to the ratio of realized triplets over all possible triplets, including those whose every or some edges belong to interlayers, normalized by norm_factor. If norm_factor == :max, then the ratio is normalized by maximum(mg.adjacency_tensor). This function overrides Graphs.jl's global_clustering_coefficient, since the latter does not consider cliques where two nodes are the same node but in different layers/interlayers. See De Domenico et al. (2013).

MultilayerGraphs.multilayer_kronecker_delta — Method

multilayer_kronecker_delta(dims...)

Returns a 4 dimensional kronecker delta with size equal to dims.

MultilayerGraphs.multilayer_weighted_global_clustering_coefficient — Method

multilayer_weighted_global_clustering_coefficient(mg::M, norm_factor::Union{Float64, Symbol} = :max) where {M <: AbstractMultilayerGraph}

Return the complete multilayer global clustering coefficient, equal to the ratio of realized triplets over all possible triplets, including those whose every or some edges belong to interlayers, normalized by norm_factor. Each triplets contributes for w[1] if all of its vertices are in one layer, w[2] if its vertices span two layers, and w[3] if they span 3 layers. If norm_factor == :max, then the ratio is normalized by maximum(mg.adjacency_tensor). This function overrides Graphs.jl's global_clustering_coefficient, since the latter does not consider cliques where two nodes are the same node but in different layers/interlayers. See De Domenico et al. (2013).

MultilayerGraphs.multiplex_interlayer — Method

multiplex_interlayer(nv::Int64, name::Symbol,layer_1::Symbol, layer_2::Symbol, graph_type::Type{G}; forbidden_vertices::Vector{MultilayerVertex{T}}, forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}) where {T <: Union{ <: Integer, AbstractVertex}, G <: AbstractGraph{T};  IsWeighted{G}}

Return an Interlayer{T,U,G} that has edges only between vertices that represent the same node.

ARGUMENTS

nv::Int64: number of vertices;

name::Symbol: name of the Interlayer;

layer_1::Symbol: The first Layer it connects;

layer_2::Symbol: The second Layer it connects;

graph_type::Type{G}: the underlying graph type;

U::Type = Float64: the eltype of the adjacency matrix of the Interlayer. Note: it doesn't have to coincide with the underlying graph's adjacency matrix eltype, since right now there is no guarantee that all Graphs.jl's extension will allow the user to set such eltype on new graph types.

forbidden_vertices::Vector{MultilayerVertex{T}}: list of vertices that are not considered present in the Interlayer;

forbidden_edges::Vector{NTuple{2, MultilayerVertex{T}}}: list of edges whose existence is a priori excluded from the Interlayer.

MultilayerGraphs.nIn — Method

nIn(mg::M) where {M <: AbstractMultilayerGraph }

Return the number of interlayers in mg.

MultilayerGraphs.nl — Method

nl(mg::M) where {M <: AbstractMultilayerGraph }

Return the number of layers in mg.

MultilayerGraphs.nn — Method

nn(mg::M) where {M <: AbstractMultilayerGraph }

Return the number of nodes in mg.

MultilayerGraphs.nodes — Method

nodes(mg::M) where {M <: AbstractMultilayerGraph}

Return the nodes of the AbstractMultilayerGraph mg.

MultilayerGraphs.overlay_clustering_coefficient — Method

overlay_clustering_coefficient(mg::M, norm_factor::Union{Float64, Symbol} = :max) where {M <: AbstractMultilayerGraph}

Return the overlay clustering coefficient as calculated in De Domenico et al. (2013).

MultilayerGraphs.specify_interlayer! — Method

specify_interlayer!(mg::M, new_interlayer::In; symmetric_interlayer_name::Symbol) where { T, U, G<: AbstractGraph{T}, M <: MultilayerDiGraph{T, U}, In <: Interlayer{T,G}; IsDirected{M}}

Specify the interlayer new_interlayer as part of mg. The underlying graph of new_interlayer must be directed.

MultilayerGraphs.specify_interlayer! — Method

specify_interlayer!(mg::M, new_interlayer::In; symmetric_interlayer_name::String) where { T, U, G<: AbstractGraph{T}, M <: MultilayerGraph{T, U}, In <: Interlayer{T,U,G}}

Specify the interlayer new_interlayer as part of mg. The underlying graph of new_interlayer must be undirected.

MultilayerGraphs.specify_interlayer! — Method

specify_interlayer!(mg::M, layer_1::Symbol, layer_2::Symbol, graph::G; new_interlayer_name::Symbol, symmetric_interlayer_name::Symbol ,  forbidden_vertices::Tuple{Vararg{MultilayerVertex{T}}}, forbidden_edges::Tuple{Vararg{NTuple{2, MultilayerVertex{T}}}} ) where { T, U, G<: AbstractGraph{T}, M <: MultilayerDiGraph{T,U}}

Specify an interlayer which is represented by graph between layer_1 and layer_2. The Interlayer's name can be specified via new_interlayer_name, as the name of its corresponding Interlayer between layer_2 and layer_1 (via the symmetric_interlayer_name) which is the same graph just with the adjacency matrix reordered.

MultilayerGraphs.src — Method

src(e::AbstractMultilayerEdge)

MultilayerGraphs.tensoreig — Method

tensoreig(A, a, b; chis=nothing, eps=0,
          return_error=false, print_error=false,
          break_degenerate=false, degeneracy_eps=1e-6,
          norm_type=:frobenius, hermitian=false)

Finds the "right" eigenvectors and eigenvalues of A. The indices of A are permuted so that the indices listed in the Array/Tuple a are on the "left" side and indices listed in b are on the "right". The resulting tensor is then reshaped to a matrix, and eig is called on this matrix to get the vector of eigenvalues E and matrix of eigenvectors U. Finally, U is reshaped to a tensor that has as its last index the one that enumerates the eigenvectors and the indices in a as its first indices.

Truncation and error printing work as with tensorsvd.

Note that no iterative techniques are used, which means that choosing to truncate provides no performance benefits: All the eigenvalues are computed in any case.

The keyword argument hermitian (false by default) tells the algorithm whether the reshaped matrix is Hermitian or not. If hermitian=true, then A = Udiagm(E)U' up to the truncation error.

Output is E, U, and possibly error, if return_error=true. Here E is a vector of eigenvalues values and U[:,...,:,k] is the kth eigenvector.

MultilayerGraphs.tensorsplit — Method

tensorsplit(A, a, b; kwargs...)

Calls tensorsvd with the arguments given to it to decompose the given tensor A with indices a on one side and indices b on the other. It then splits the diagonal matrix of singular values into two with a square root and multiplies these weights into the isometric tensors. Thus tensorsplit ends up splitting A into two parts, which are then returned, possibly together with auxiliary data such as a truncation error. If the keyword argument hermitian=true, an eigenvalue decomposition is used in stead of an SVD. All the keyword arguments are passed to either tensorsvd or tensoreig.

See tensorsvd and tensoreig for further documentation.

MultilayerGraphs.tensorsvd — Method

tensorsvd(A, a, b;
          chis=nothing, eps=0,
          return_error=false, print_error=false,
          break_degenerate=false, degeneracy_eps=1e-6,
          norm_type=:frobenius)

Singular valued decomposes a tensor A. The indices of A are permuted so that the indices listed in the Array/Tuple a are on the "left" side and indices listed in b are on the "right". The resulting tensor is then reshaped to a matrix, and this matrix is SVDed into Udiagm(S)Vt. Finally, the unitary matrices U and Vt are reshaped to tensors so that they have a new index coming from the SVD, for U as the last index and for Vt as the first, and U has indices a as its first indices and V has indices b as its last indices.

If eps>0 then the SVD may be truncated if the relative error can be kept below eps. For this purpose different dimensions to truncate to can be tried, and these dimensions should be listed in chis. If chis is nothing (the default) then the full range of possible dimensions is tried. If breakdegenerate=false (the default) then the truncation never cuts between degenerate singular values. degeneracyeps controls how close the values need to be to be considered degenerate.

norm_type specifies the norm used to measure the error. This defaults to :frobenius, which means that the error measured is the Frobenius norm of the difference between A and the decomposition, divided by the Frobenius norm of A. This is the same thing as the 2-norm of the singular values that are truncated out, divided by the 2-norm of all the singular values. The other option is :trace, in which case a 1-norm is used instead.

If print_error=true the truncation error is printed. The default is false.

If return_error=true then the truncation error is also returned. The default is false.

Note that no iterative techniques are used, which means choosing to truncate provides no performance benefits: The full SVD is computed in any case.

Output is U, S, Vt, and possibly error. Here S is a vector of singular values and U and Vt are isometric tensors (unitary if the matrix that is SVDed is square and there is no truncation) such that Udiag(S)Vt = A, up to truncation errors.

MultilayerGraphs.to_matrix — Method

Transpose A so that the indices listed in a are on the left and the indices listed in b on the right, and reshape A into a matrix.

a and b should be arrays of Integers that together include the numbers from 1 to ndims(A). Alternatively they can be just individual Integers, if only one index is left on one side.

If returntensorshape is true (by default it's not) return, in addition to the matrix, the shape of the tensor after the transpose but before the reshape.

MultilayerGraphs.von_neumann_entropy — Method

von_neumann_entropy(mg::M) where {T,U,  M <: MultilayerGraph{T, U}}

Compute the Von Neumann entropy of mg, according to De Domenico et al. (2013). Only for undirected multilayer graphs.

MultilayerGraphs.weight — Method

weight(e::AbstractMultilayerEdge)